requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
attrs==25.3.0
    # via zeep
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.2
    # via pyjwt
idna==3.10
    # via requests
isodate==0.7.2
    # via zeep
lxml==5.3.2
    # via zeep
more-itertools==10.6.0
    # via simple-salesforce
numpy==2.2.4
    # via pandas
pandas==2.2.3
    # via -r requirements.in
platformdirs==4.3.7
    # via zeep
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.10.1
    # via simple-salesforce
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via
    #   pandas
    #   zeep
requests==2.32.3
    # via
    #   -r requirements.in
    #   requests-file
    #   requests-toolbelt
    #   simple-salesforce
    #   zeep
requests-file==2.1.0
    # via zeep
requests-toolbelt==1.0.0
    # via zeep
simple-salesforce==1.12.6
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
typing-extensions==4.13.2
    # via simple-salesforce
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
zeep==4.3.1
    # via simple-salesforce
